The scintilla editor now allows rectangular selection/multi line typing. To use it, hold ALT to select and then start typing, like in SSMS.
Fixed a bug where the direct x settings change message was popping up when the settings window was loading and the user had not actually changed it.

Renabled key tips for the ribbon, but for now they can only be triggered by F10.  Alt has to stay reserved for Scintilla!