Finished implementing dock layout panel mode for the results grids. Printing and exporting also work.  There may be possible bugs we we can eliminate them later. We may also want to add the ability to have a single grid show up full sized. I will add an enhancement for that one.
Also added ability to switch which layout to use via the settings menu. Note that toggling the setting requires the whole query pane to be reconstructed. There really isn't a good way around this right now, but I think it is possible to have it switch.  I will look into that later.